reorganization, first steps
1. Single project version definition in the CMake main file.
2. Generation of fluid_config.h and version.h by cmake.
3. Fixed generation of fluidlite.pc in case of absolute install paths.
4. Renaming the headers' directory to 'fluidlite/', so the project can
be installed system wide without conflicts with the fluidsynth project.
4. Single compilation of all sources, using a cmake object library. The
objects are linked to the shared library and archived to the static one.
