+++ fluidlite.pc.in
@@ -1,10 +1,10 @@
 prefix=@CMAKE_INSTALL_PREFIX@
 exec_prefix=${prefix}
-libdir=${exec_prefix}/lib
-includedir=${prefix}/include
+libdir=@fluidlite_libdir@
+includedir=@fluidlite_includedir@
 
 Name: fluidlite
 Description: Software SoundFont synth
-Version: 1.2.1
+Version: @fluidlite_VERSION@
 Libs: -L${libdir} -lfluidlite @ADDITIONAL_LIBS@
 Cflags: -I${includedir}
