+++ include/fluidlite/misc.h
