+++ src/fluid_config.cmake
@@ -1,11 +1,11 @@
-/* src/config.h.in.  Generated from configure.ac by autoheader.  */
+/* fluid_config.h will be generated by cmake from src/fluid_config.cmake */
 
 
 /* Define to activate debugging message */
 #undef DEBUG
 
 /* Version number of package */
-#define VERSION "1.0.9"
+#define VERSION "@fluidlite_VERSION@"
 
 /* Define to 1 if your processor stores words with the most significant byte
    first (like Motorola and SPARC, unlike Intel and VAX). */
